python-backend/requirements.txt
pillow~=11.1.0
tensorflow~=2.18.0
numpy~=2.0.2
Flask~=3.1.0
matplotlib~=3.10.0
pandas~=2.2.3
scikit-learn~=1.6.1
kagglehub~=0.3.6